Rework sRGB/Linear color workflow
Maybe this is better? The idea is that textures are typically
interpreted as sRGB on load, then all operations and renders are linear,
then finally the very last rendering to the screen converts to sRGB
again.

In order to not completely crush colors, I found this only works well
with float rendertargets. Maybe that is a symptom of things not being
implemented properly? Or maybe that's just expected behavior? This is
probably not the last time we tweak color management in renin.

See https://threejs.org/docs/#manual/en/introduction/Color-management
for more information about color management in Three.js.