Zoom on mouse cursor instead of playhead
Issue #40 describes unwanted behavior of zooming on the center of the
zoomed screen rather than on the global center of the timeline. I
believe the actual issue of unwanted zooming to be the fact that zooming
is always oriented around the playhead, which is often confusing.

This commit updates zooming to be oriented around the mouse cursor
(since you use the mouse to zoom). Except for some rounding errors,
which makes it perhaps not Perfect™, I think this feels quite natural,
and therefore probably also fixes #40.