Add some metadata to npm publishing
It would be nice to include readme and license as well, but readme will
be copied along automatically if we as part of the release step copy the
top-level readme into the published folder (or add a secondary readme
inside with release specific information).

And the license is just still missing, so I'll leave that for later.