Repair stepOffset in sync.flash
I couldn't find any traces of us actually having used stepOffset before,
so I can't say for sure what the intended purpose was, but this makes it
behave more how I was expecting it, letting you offset when in a
beat/bar flashes occur.

The trick is that we need the value for startStep to change exactly when
we reach the step of its new value. That was out of sync without this
fix, giving strange results.