Example project: Refer to sibling renin by default
For some reason, the absence of this, has yet again this year, resulted
in an ancient renin from registry.yarnpkg.com being picked up. I mean,
if we encouraged by running with just refferingto renin in package.json
and that was it, I'd get it. But since we require renin being in a sibling
folder anyway, this is just a source of pain. And this should fix it. For
now.